% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{auc}
\alias{auc}
\title{auc}
\usage{
auc(x, y = NULL)
}
\arguments{
\item{x}{if x is a dataframe represent FPR and TPR values but if it is a vector represent FPR value.}

\item{y}{if x is a vector, y represent TPR values}
}
\value{
auc value
}
\description{
auc
}
\examples{
random.x <- sample.int(3000,1000,replace=FALSE)
class <- c(rep(TRUE, 500), rep(FALSE, 500))
pos <- sample.int(1000, 1000)
random.y <- class[order(pos)]
roc.curve <- roc.curve(random.x, random.y)
auc <- auc(roc.curve$FPR , roc.curve$TPR )
}
