% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{conditional.entropy}
\alias{conditional.entropy}
\title{Conditional Entropy}
\usage{
conditional.entropy(x, y = NULL, file = "None")
}
\arguments{
\item{x}{a vector, matrix, or data.frame.}

\item{y}{a vector. y must be NULL if x is a matrix or a data.frame}

\item{file}{In case of x been a data.frame or matrix, represent the name of output file for saving conditional entropy plot}
}
\value{
In case of x been a data.frame or matrix, conditional entropy of all variables/columns in x. Otherwise, conditional entropy of x and y
}
\description{
Conditional Entropy
}
\examples{
df <- data.frame("V1" = sample.int(2,5,replace=TRUE), "V2" =sample.int(2,5,replace=TRUE) , "V3" = sample.int(2,5,replace=TRUE), "V4" = sample.int(2,5,replace=TRUE),  "V5" = sample.int(2,5,replace=TRUE), "V6" = sample.int(2,5,replace=TRUE), "V7" = sample.int(2,5,replace=TRUE), "V8" = sample.int(2,5,replace=TRUE))
x <- sample.int(2, 5,replace=TRUE)
y <- sample.int(2,5,replace=TRUE)
conditional.entropy(df, file ="conditionalEntropy")
conditional.entropy(x = x, y = y)

}
