% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\title{Display the results of the data frame in a nice table format}
\usage{
display(x, digits = 3)
}
\arguments{
\item{x}{data.frame}

\item{digits}{numeric}
}
\value{

}
\description{
It's a simple wrapper function of the package \code{kableExtra}.
}
\examples{
#Not run:
#x = data.frame(a = seq(0, 10, 1), b = seq(1, 11, 1))
#display(x)
}
