% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_algorithm.R
\name{e.inner}
\alias{e.inner}
\title{calculate shift vector on manifold origin}
\usage{
e.inner(n, grp)
}
\arguments{
\item{n}{Numeric of origin.}

\item{grp}{Graph of manifold.}
}
\description{
calculate shift vector on manifold origin
}
