% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(x, margin = 2L)
}
\arguments{
\item{x}{a numeric vector, matrix, or data.frame.}

\item{margin}{In case of matrix or dataframe apply entropy by columns margin = 2L or by rows margin = 1L. Default 2.}
}
\value{
the sample entropy of the input.
}
\description{
Entropy
}
\examples{
col1 <-  c('False', 'False', 'True', 'True', 'True', 'True', 'False', 'True', 'False', 'False')
col2 <- c('True', 'False', 'False', 'False', 'True', 'True', 'False', 'False', 'False', 'False')
col3 <- c('a', 'b', 'c', 'a', 'a', 'a', 'l', 'a', 'l', 'a')
df <- data.frame("V1" = col1, "V2" = col2, "V3" = col3)
entropy(df)
}
