% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ar1_sym.R
\name{estimate_ar1_sym}
\alias{estimate_ar1_sym}
\title{Estimate AR1 symmetric classification error estimator}
\usage{
estimate_ar1_sym(
  st3_observed = NULL,
  st2_observed = NULL,
  st1_observed = NULL,
  weights = NULL,
  model_name = "SME",
  init_params = NULL,
  global_est = FALSE,
  global_specification = NULL,
  verbose = FALSE
)
}
\arguments{
\item{st3_observed}{numeric vector: observed status in t3}

\item{st2_observed}{numeric vector: observed status in t2}

\item{st1_observed}{numeric vector: observed status in t1}

\item{weights}{numeric: survey weights. Default is NULL, giving
equal weight to each observation.}

\item{model_name}{character: used to identify different estimated results}

\item{init_params}{named atomic vector: initial values for parameters}

\item{global_est}{logical: should a global estimator be used to determine
the starting values for local estimation}

\item{global_specification}{list: parameter bounds for global estimation}

\item{verbose}{logical: should estimation info be printed to console}
}
\value{
list
}
\description{
Estimate AR1 symmetric classification error estimator
}
