% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ar2.R
\name{estimate_ar2}
\alias{estimate_ar2}
\title{Estimate the AR(2) misclassification error model}
\usage{
estimate_ar2(
  st4_observed = NULL,
  st3_observed = NULL,
  st2_observed = NULL,
  st1_observed = NULL,
  weights = NULL,
  model_name = "SOME",
  init_params = NULL,
  global_est = FALSE,
  global_specification = NULL,
  verbose = FALSE
)
}
\arguments{
\item{st4_observed}{vector of observed binary status in the 4th wave - either logical or binary 0-1}

\item{st3_observed}{vector of observed binary status in the 3rd wave - either logical or binary 0-1}

\item{st2_observed}{vector of observed binary status in the 2nd wave - either logical or binary 0-1}

\item{st1_observed}{vector of observed binary status in the 1st wave - either logical or binary 0-1}

\item{weights}{vector of weights summing to population. NULL (default) receives equal weights}

\item{model_name}{name of the model}

\item{init_params}{initial parameters for the optimization}

\item{global_est}{logical indicating whether to estimate the model globally first, then
use results as starting values for the local optimization}

\item{global_specification}{list of bounds for the global optimization}

\item{verbose}{logical indicating whether to print information as the estimation goes}
}
\value{
list of estimated parameters and results
}
\description{
Estimate the AR(2) misclassification error model
}
