% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_evaluate.R
\name{evaluateEmbedding}
\alias{evaluateEmbedding}
\title{Calculates evaluate measures of the embedding.}
\usage{
evaluateEmbedding(Y, X, k = 12)
}
\arguments{
\item{Y}{after embedding data or dist.}

\item{X}{pre embedding data or dist.}

\item{k}{number of neighbors for evaluate measures.}
}
\value{
evaluate measures.
}
\description{
Calculates evaluate measures of the embedding.
}
