% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sme_evaluate.R
\name{evaluateSme}
\alias{evaluateSme}
\title{Compare the embedding of sme with the conventional methods.}
\usage{
evaluateSme(sme, k = 12, embk = 10, return.embs = F, plot.only = F,
  plot.fig = T, col = NULL)
}
\arguments{
\item{sme}{self-manifolding embedding object.}

\item{k}{number of neighbors for Trustworthiness and Continuity.}

\item{embk}{number of neighbors for embedding of LLE and Isomap.}

\item{return.embs}{if\code{T}, returns embedding results with listed.}

\item{plot.only}{if \code{T}, doesn't calculate evaluation.}

\item{plot.fig}{if \code{T}, plot embedding result.}

\item{col}{color of pca.}
}
\value{
a matrix contains evaluation.
}
\description{
Compare the embedding of sme with the conventional methods.
}
