% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Example data set}
\format{
## `ex_data`
A data frame with 12651 rows and 4 columns:
\describe{
  \item{status}{Status: either TRUE or FALSE}
  \item{weight}{Survey weight}
  \item{wave}{Wave: either 1, 2, or 3, giving the periiod}
  ...
}
}
\source{
self-created
}
\usage{
ex_data
}
\description{
Example data set
}
\keyword{datasets}
