% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar1_tenure.R
\name{ex_gaussian_density}
\alias{ex_gaussian_density}
\title{Density of exGaussian distribution}
\usage{
ex_gaussian_density(x, sigma, lambda)
}
\arguments{
\item{x}{numeric value}

\item{sigma}{numeric: standard deviation for gaussian component}

\item{lambda}{numeric: parameter of exponential component}
}
\value{
numeric atomic vector length 1
}
\description{
Sum of a gaussian and exponential random variable.
Former distributed N(0,sigma), latter distributed exp(lambda)
}
\keyword{internal}
