% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{filter}
\alias{filter}
\title{Metric filter}
\usage{
filter(
  df,
  class,
  by = "AUC",
  uplimit = .Machine$integer.max,
  lowlimit = -.Machine$integer.max
)
}
\arguments{
\item{df}{data.frame. Each colum represent a variable}

\item{class}{name of the class variable}

\item{by}{metric to apply the filter. AUC, Variance and Entropy are available.}

\item{uplimit}{metric uplimit value}

\item{lowlimit}{metric uplimit value. Default}
}
\value{
Filtered data.frame
}
\description{
Metric filter
}
\examples{
class <- c(rep(TRUE, 50), rep(FALSE, 50))
pos <- sample.int(100, 100)
random.y <- class[order(pos)]
df.numeric <- data.frame("V1" = sample.int(10,100,replace=TRUE), "V2" = sample.int(10,100,replace=TRUE), "V3" = sample.int(100,100,replace=TRUE), "Class" =  random.y)
df.cat <- data.frame("V1" = as.factor(sample.int(3,100,replace=TRUE)), "V2" =  as.factor(sample.int(3,100,replace=TRUE)),  "V4" =  as.factor(sample.int(3,100,replace=TRUE)), "Class" =  random.y)
filter(df.numeric, class = "Class", by = "AUC", uplimit = 1, lowlimit = 0.5)
filter(df.numeric, class = "Class", by = "Variance", uplimit = 100, lowlimit = 0)
filter(df.cat, class = "Class", by = "Entropy", uplimit = 2, lowlimit = 0)
filter(df.cat, class = "Class", by = "Entropy")
}
