% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sme.R
\name{fn_Joint_Prob}
\alias{fn_Joint_Prob}
\title{Joint probability of t1 and t2 observed and true statuses}
\usage{
fn_Joint_Prob(S_Q2, S_Q2star, S_Q1, S_Q1star, theta_01, theta_02, pi, mu)
}
\arguments{
\item{S_Q2}{numeric: 0 or 1 - observed status in period t2}

\item{S_Q2star}{numeric: 0 or 1 - true status in period t2}

\item{S_Q1}{numeric: 0 or 1 - observed status in period t1}

\item{S_Q1star}{numeric: 0 or 1 - true status in period t1}

\item{theta_01}{numeric: parameter of true transition process specifying
the probability (after applying probit) exiting status 1}

\item{theta_02}{numeric: parameter of true transition process specifying
the probability (after applying probit) of entering status 1}

\item{pi}{numeric: parameter of noise process. The probability of
misclassification is equal to 1-Phi(pi), where Phi() is the probit
transformation}

\item{mu}{numeric: parameter of true process. Specifies
unconditional distribution in t1, where P(S_Q1star = 1) = mu}
}
\value{
numeric double: joint probability between 0 and 1
}
\description{
This function gives the joint probability that some specified combination
of true and observed statuses (S) are experienced over to periods. The
This function requires specifying parameter values.
}
