% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar1_sym.R
\name{get_ar1_sym_joint_probability}
\alias{get_ar1_sym_joint_probability}
\title{Joint probability of t1 and t2 observed and true statuses}
\usage{
get_ar1_sym_joint_probability(
  st2_observed,
  st2_true,
  st1_observed,
  st1_true,
  theta_01,
  theta_02,
  err,
  mu
)
}
\arguments{
\item{st2_observed}{numeric: 0 or 1 - observed status in period t2}

\item{st2_true}{numeric: 0 or 1 - true status in period t2}

\item{st1_observed}{numeric: 0 or 1 - observed status in period t1}

\item{st1_true}{numeric: 0 or 1 - true status in period t1}

\item{theta_01}{numeric: parameter of true transition process specifying
the probability (after applying probit) exiting status 1}

\item{theta_02}{numeric: parameter of true transition process specifying
the probability (after applying probit) of entering status 1}

\item{err}{numeric: parameter of noise process. The probability of
misclassification is equal to 1-Phi(err), where Phi() is the probit
transformation}

\item{mu}{numeric: parameter of true process. Specifies
unconditional distribution in t1, where P(st1_true = 1) = mu}
}
\value{
numeric double: joint probability between 0 and 1
}
\description{
This function gives the joint probability that some specified combination
of true and observed statuses (S) are experienced over to periods. The
This function requires specifying parameter values.
}
\keyword{internal}
