% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar1_tenure_two_sigma.R
\name{get_ar1_tenure_full_likelihood_two_sigma}
\alias{get_ar1_tenure_full_likelihood_two_sigma}
\title{Log-likelihood of full observed sample - tenure/unemployment duration}
\usage{
get_ar1_tenure_full_likelihood_two_sigma(
  st1_observed,
  st2_observed,
  st3_observed,
  g_1,
  g_2,
  g_3,
  h_1,
  h_2,
  h_3,
  theta_1,
  theta_2,
  sigma_g,
  sigma_h,
  lambda_g,
  lambda_h,
  err,
  skip_checks = FALSE
)
}
\arguments{
\item{st1_observed}{numeric: 0 or 1 - observed status in period t1}

\item{st2_observed}{numeric: 0 or 1 - observed status in period t2}

\item{st3_observed}{numeric vector: observed status in t3}

\item{g_1}{numeric: observed tenure t1}

\item{g_2}{numeric: observed tenure t2}

\item{g_3}{numeric: observed tenure t3}

\item{h_1}{numeric: observed unemployment duration t1}

\item{h_2}{numeric: observed unemployment duration t2}

\item{h_3}{numeric: observed unemployment duration t3}

\item{theta_1}{numeric: parameter of true transition process specifying
the probability (after applying probit) exiting status 1}

\item{theta_2}{numeric: parameter of true transition process specifying
the probability (after applying probit) of entering status 1}

\item{lambda_g}{numeric: parameter of exponentially distributed
tenure}

\item{lambda_h}{numeric: parameter of exponentially distributed
unemployment duration}

\item{err}{numeric: parameter of noise process. The probability of
misclassification is equal to 1-Phi(err), where Phi() is the probit
transformation}
}
\value{
numeric: log-likelihood
}
\description{
Log-likelihood of full observed sample - tenure/unemployment duration
}
