% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar1_tenure_two_sigma_with_mu.R
\name{get_ar1_tenure_individual_likelihood_two_sigma_with_mu}
\alias{get_ar1_tenure_individual_likelihood_two_sigma_with_mu}
\title{Likelihood vector - tenure and unemployment duration}
\usage{
get_ar1_tenure_individual_likelihood_two_sigma_with_mu(
  st1_observed,
  st2_observed,
  st3_observed,
  g_1,
  g_2,
  g_3,
  h_1,
  h_2,
  h_3,
  theta_1,
  theta_2,
  sigma_g,
  sigma_h,
  lambda_g,
  lambda_h,
  err,
  mu,
  by_true_status = FALSE,
  skip_checks = FALSE
)
}
\arguments{
\item{st1_observed}{numeric: 0 or 1 - observed status in period t1}

\item{st2_observed}{numeric: 0 or 1 - observed status in period t2}

\item{g_1}{numeric: observed tenure t1}

\item{g_2}{numeric: observed tenure t2}

\item{g_3}{numeric: observed tenure t3}

\item{h_1}{numeric: observed unemployment duration t1}

\item{h_2}{numeric: observed unemployment duration t2}

\item{h_3}{numeric: observed unemployment duration t3}

\item{theta_1}{numeric: parameter of true transition process specifying
the probability (after applying probit) exiting status 1}

\item{theta_2}{numeric: parameter of true transition process specifying
the probability (after applying probit) of entering status 1}

\item{lambda_g}{numeric: parameter of exponentially distributed
tenure}

\item{lambda_h}{numeric: parameter of exponentially distributed
unemployment duration}

\item{err}{numeric: parameter of noise process. The probability of
misclassification is equal to 1-Phi(err), where Phi() is the probit
transformation}

\item{mu}{numeric: parameter of true process. Specifies
unconditional distribution in t1, where P(st1_true = 1) = mu}

\item{by_true_status}{logical: TRUE breaks the likelihood down by each of the
8 combinations of true statuses over three periods while FALSE integrates over
the true statuses, giving a single likelihood for each element}
}
\value{
numeric data frame with 1 row if `by_true_status = FALSE` else 8 rows,
and columns equal to the length of the input vectors
}
\description{
Vectorized version of [get_ar1_tenure_joint_probability], but integrating
out the true statuses. Gives the
}
\examples{
get_ar1_tenure_individual_likelihood(
st1_observed = 1,
st2_observed = 1,
st3_observed = 1,
g_1          = 0.5,
g_2          = 0.75,
g_3          = 1.1,
h_1          = 0,
h_2          = 0,
h_3          = 0,
err          = 1.65,
mu           = 0.1,
theta_1      = 1.65,
theta_2      = -1.65,
sigma        = 0.1,
lambda_h     = 2,
lambda_g     = 2
)
}
