% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar1_tenure.R
\name{get_ar1_tenure_joint_probability}
\alias{get_ar1_tenure_joint_probability}
\title{Joint probability - tenure and unemployment duration}
\usage{
get_ar1_tenure_joint_probability(
  st1_observed,
  st2_observed,
  st3_observed,
  st1_true,
  st2_true,
  st3_true,
  g_1,
  g_2,
  g_3,
  h_1,
  h_2,
  h_3,
  err,
  mu,
  theta_1,
  theta_2,
  sigma,
  lambda_h,
  lambda_g
)
}
\arguments{
\item{st1_observed}{numeric: 0 or 1 - observed status in period t1}

\item{st2_observed}{numeric: 0 or 1 - observed status in period t2}

\item{st1_true}{numeric: 0 or 1 - true status in period t1}

\item{st2_true}{numeric: 0 or 1 - true status in period t2}

\item{g_1}{numeric: observed tenure t1}

\item{g_2}{numeric: observed tenure t2}

\item{g_3}{numeric: observed tenure t3}

\item{h_1}{numeric: observed unemployment duration t1}

\item{h_2}{numeric: observed unemployment duration t2}

\item{h_3}{numeric: observed unemployment duration t3}

\item{err}{numeric: parameter of noise process. The probability of
misclassification is equal to 1-Phi(err), where Phi() is the probit
transformation}

\item{mu}{numeric: parameter of true process. Specifies
unconditional distribution in t1, where P(st1_true = 1) = mu}

\item{sigma}{numeric: standard deviation parameter of normally
distributed measurement error term}

\item{lambda_h}{numeric: parameter of exponentially distributed
unemployment duration}

\item{lambda_g}{numeric: parameter of exponentially distributed
tenure}
}
\value{
numeric: joint probability
}
\description{
This function gives the full joint probability of a single individual
(i.e. non-vectorized) having the given observed potentially
misclassified status (s),
true status (s_star), observed potentially mismeasured or
misclassified tenure (g), observed potentially mismeasured or misclassified
unemployment duration. This probability is conditioned on the
parameters -- mu, theta_1, theta_2, sigma, lambda_h, lambda_g --
these specify the underlying distributions.
}
