% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar1_tenure_mult_sigma.R
\name{get_ar1_tenure_joint_probability_mult_sigma}
\alias{get_ar1_tenure_joint_probability_mult_sigma}
\title{Joint probability - tenure and unemployment duration}
\usage{
get_ar1_tenure_joint_probability_mult_sigma(
  st1_observed,
  st2_observed,
  st3_observed,
  st1_true,
  st2_true,
  st3_true,
  g_1,
  g_2,
  g_3,
  h_1,
  h_2,
  h_3,
  theta_1,
  theta_2,
  sigma_g0,
  sigma_g1,
  sigma_h0,
  sigma_h1,
  lambda_g,
  lambda_h,
  err,
  skip_checks = FALSE
)
}
\arguments{
\item{st1_observed}{numeric: 0 or 1 - observed status in period t1}

\item{st2_observed}{numeric: 0 or 1 - observed status in period t2}

\item{st1_true}{numeric: 0 or 1 - true status in period t1}

\item{st2_true}{numeric: 0 or 1 - true status in period t2}

\item{g_1}{numeric: observed tenure t1}

\item{g_2}{numeric: observed tenure t2}

\item{g_3}{numeric: observed tenure t3}

\item{h_1}{numeric: observed unemployment duration t1}

\item{h_2}{numeric: observed unemployment duration t2}

\item{h_3}{numeric: observed unemployment duration t3}

\item{theta_1}{numeric: parameter of true transition process specifying
the probability (after applying probit) exiting status 1}

\item{theta_2}{numeric: parameter of true transition process specifying
the probability (after applying probit) of entering status 1}

\item{lambda_g}{numeric: parameter of exponentially distributed
tenure}

\item{lambda_h}{numeric: parameter of exponentially distributed
unemployment duration}

\item{err}{numeric: parameter of noise process. The probability of
misclassification is equal to 1-Phi(err), where Phi() is the probit
transformation}

\item{sigma}{numeric: standard deviation parameter of normally
distributed measurement error term}
}
\value{
numeric: joint probability
}
\description{
This function gives the full joint probability of a single individual
(i.e. non-vectorized) having the given observed potentially
misclassified status (s),
true status (s_star), observed potentially mismeasured or
misclassified tenure (g), observed potentially mismeasured or misclassified
unemployment duration. This probability is conditioned on the
parameters -- mu, theta_1, theta_2, sigma, lambda_h, lambda_g --
these specify the underlying distributions.
}
\examples{
get_ar1_tenure_joint_probability(
st1_observed = 1,
st2_observed = 1,
st3_observed = 1,
st1_true     = 1,
st2_true     = 1,
st3_true     = 1,
g_1          = 0.5,
g_2          = 0.75,
g_3          = 1.1,
h_1          = 0,
h_2          = 0,
h_3          = 0,
err          = 1.65,
mu           = 0.1,
theta_1      = 1.65,
theta_2      = -1.65,
sigma        = 0.1,
lambda_h     = 2,
lambda_g     = 2
)
}
