% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar2_log_expected_likelihood.R
\name{get_ar2_expected_log_likelihood}
\alias{get_ar2_expected_log_likelihood}
\title{AR(2) misclassification expected log likelihood}
\usage{
get_ar2_expected_log_likelihood(
  s4,
  s3,
  s2,
  s1,
  weights = NULL,
  theta11,
  theta12,
  theta21,
  theta22,
  err,
  mu
)
}
\arguments{
\item{s4}{Binary 0-1 **observed** status vector in the 4th wave}

\item{s3}{Binary 0-1 **observed** status vector in the 4th wave}

\item{s2}{Binary 0-1 **observed** status vector in the 4th wave}

\item{s1}{Binary 0-1 **observed** status vector in the 4th wave}

\item{weights}{vector of weights summing to population. NULL (default) receives equal weights}

\item{err}{Misclassification probability - scalar between 0 and 1}
}
\value{
numeric
}
\description{
Used as the E-step in the EM algorithm.
}
