% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_optimize.R
\name{grpOptimize}
\alias{grpOptimize}
\title{Optimize graph to preserve the distance between previous and after the embedding.}
\usage{
grpOptimize(X, som, grp)
}
\arguments{
\item{X}{target data.}

\item{som}{learned som.}

\item{grp}{target graph.}
}
\description{
Optimize graph to preserve the distance between previous and after the embedding.
}
