% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_optimize.R
\name{grpWeighting}
\alias{grpWeighting}
\title{Objective function for grpOptimize}
\usage{
grpWeighting(t, X, X.dist, som, grp)
}
\arguments{
\item{t}{parameter.}

\item{X}{target data.}

\item{X.dist}{dist of target data.}

\item{som}{learned som.}

\item{grp}{target graph.}
}
\value{
error of distance due to the embedding with multiplying weights of the graph by t.
}
\description{
Objective function for grpOptimize
}
