% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_algorithm.R
\name{matlist}
\alias{matlist}
\title{Deforms listed list (a list nested by list) as matrix}
\usage{
matlist(lst, nrow = length(lst), ncol = length(lst[[nrow]]),
  rname = names(lst), cname = names(lst[[nrow]]))
}
\arguments{
\item{lst}{listed list.}

\item{nrow}{nrow of returned matrix.}

\item{ncol}{ncol of returned matrix.}

\item{rname}{rownames of returned matrix.}

\item{cname}{colnames of returned matrix.}
}
\value{
matrix deformed from listed list.
}
\description{
Deforms listed list (a list nested by list) as matrix
}
