% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x, margin = 1L)
}
\arguments{
\item{x}{numeric, matrix or dataframe input vector.}

\item{margin}{1 - normalize by rows and 2 - normalize by columns. Default is 1.}
}
\value{
A normalized numeric or data.frame.
}
\description{
Currently implemented for numeric vectors, numeric matrices and data.frame normalization.
}
\examples{
df <- data.frame("V1" = 1:3, "V2" = 3:5, "V3" = 5:7)
normalize(df, margin = 1L)
normalize(df, margin = 2L)

m <- matrix(c(1,2,3,3,4,5,5,6,7), ncol = 3, byrow = FALSE)
normalize(m, margin = 1L)

v <- c(1, 3, 5)
normalize(v)
}
