% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sme_algorithm.R
\name{p}
\alias{p}
\title{Calculates coordinates on manifold.}
\usage{
p(X, A, V)
}
\arguments{
\item{X}{matrix of high dimensional vectors.}

\item{A}{origin of manifold axis.}

\item{V}{direction of manifold axis.}
}
\description{
Calculates coordinates on manifold.
}
