% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar2_log_expected_likelihood.R
\name{p_full_misclass}
\alias{p_full_misclass}
\title{Probablity of observing the `s` given the
true status `strue` and the
probability of misclassification `err`}
\usage{
p_full_misclass(s4, s3, s2, s1, s4true, s3true, s2true, s1true, err)
}
\arguments{
\item{s4}{Binary 0-1 **observed** status vector in the 4th wave}

\item{s3}{Binary 0-1 **observed** status vector in the 4th wave}

\item{s2}{Binary 0-1 **observed** status vector in the 4th wave}

\item{s1}{Binary 0-1 **observed** status vector in the 4th wave}

\item{s4true}{Binary 0-1 **true** status vector in the 4th wave}

\item{s3true}{Binary 0-1 **true** status vector in the 4th wave}

\item{s2true}{Binary 0-1 **true** status vector in the 4th wave}

\item{s1true}{Binary 0-1 **true** status vector in the 4th wave}

\item{err}{Misclassification probability - scalar between 0 and 1}
}
\value{
vector of same length as `s4`
}
\description{
Probablity of observing the `s` given the
true status `strue` and the
probability of misclassification `err`
}
