% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar2_log_expected_likelihood.R
\name{p_misclass}
\alias{p_misclass}
\title{Likelihood of observing the `s` and `strue` given the probability of misclassification `err`}
\usage{
p_misclass(s, strue, err)
}
\arguments{
\item{s}{observed status}

\item{strue}{true status}

\item{err}{probability of misclassification}
}
\value{
numeric
}
\description{
Likelihood of observing the `s` and `strue` given the probability of misclassification `err`
}
