% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar2_log_expected_likelihood.R
\name{p_num_misclass}
\alias{p_num_misclass}
\title{Probability of getting `num_misclass` misclassifications when
the probability of misclassification is `err`}
\usage{
p_num_misclass(err, num_misclass, num_correct = 4 - num_misclass)
}
\arguments{
\item{err}{probability of misclassification}

\item{num_misclass}{number of misclassifications}

\item{num_correct}{number of correct classifications: typicall 4 - num_misclass}
}
\value{
numeric
}
\description{
Probability of getting `num_misclass` misclassifications when
the probability of misclassification is `err`
}
