% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ar2_log_expected_likelihood.R
\name{p_true_cond_2period}
\alias{p_true_cond_2period}
\title{Conditional probability of observing `s3true` given `s2true` and `s1true`
and parameters}
\usage{
p_true_cond_2period(
  s3true,
  s2true,
  s1true,
  theta11,
  theta12,
  theta21,
  theta22,
  mu
)
}
\arguments{
\item{mu}{Unconditional employment parameter}
}
\value{
numeric
}
\description{
Conditional probability of observing `s3true` given `s2true` and `s1true`
and parameters
}
