% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\docType{class}
\name{planesom}
\alias{planesom}
\title{Initialize plane formed SOM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
planesom
}
\arguments{
\item{dim}{dimension of the input data}

\item{r}{length of forms}

\item{c}{height of forms}

\item{weights}{initial weights}

\item{neighbor}{size of the neighborhood kernel}

\item{alpha}{learning late}

\item{sigma}{variance of the weights propagation}

\item{collect.stats}{whether to save learning history}
}
\value{
SOM object
}
\description{
Initialize plane formed SOM
}
\keyword{datasets}
