% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.df}
\alias{read.df}
\title{Read a csv file}
\usage{
read.df(
  file.path,
  sep = ";",
  max.n.factor = 3,
  header = FALSE,
  row.names = NULL,
  col.names = NULL,
  n.rows = -1L,
  transpose = FALSE
)
}
\arguments{
\item{file.path}{Name of input file}

\item{sep}{Field separator. Default ;}

\item{max.n.factor}{maximum number of column characters to be a factor. Default 3.}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line. Default FALSE.}

\item{row.names}{a vector of row names. Default NULL.}

\item{col.names}{a vector of col names. Default NULL.}

\item{n.rows}{integer: the maximum number of rows to read in. In case of Header = TRUE, it does not count for the number of read lines.}

\item{transpose}{If TRUE, transpose the result. Default FALSE.}
}
\value{
A data frame containing a representation of the data in the file.
}
\description{
Read a csv file
}
\examples{
df = data.frame("V1" = 1:5, "V2" = c("a", "b", "c", "d", "e"), "V3" = c("alto", "bajo", "bajo", "alto", "alto"))
write.df(df, "bench1.txt")
read.df(file.path = "bench1.txt")

write.df(df, "bench2.txt", header = TRUE)
read.df(file.path = "bench2.txt", header = TRUE)

write.df(df, "bench2.txt", append = TRUE)
read.df(file.path = "bench2.txt", header = FALSE)

write.df(df, "bench2.txt", sep = ";", header = FALSE, append = TRUE)
read.df(file.path = "bench2.txt", sep = ";", header = TRUE,  max.n.factor = 3, row.names = c('X1', 'X2', 'X3', 'X4', 'X5'), col.names = c('A', 'B', 'C'), n.rows = 5, transpose = FALSE)


}
