% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{roc.curve}
\alias{roc.curve}
\title{Roc curve}
\usage{
roc.curve(x, y, file = "None", AUC = FALSE)
}
\arguments{
\item{x}{a numerical vector}

\item{y}{represent logical class TRUE/FALSE.}

\item{file}{name of output file for saving roc curve plot.}

\item{AUC}{return auc value AUC = TRUE}
}
\value{
Return roc curve values: TPR, FPR or AUC
}
\description{
Roc curve
}
\examples{
random.x <- sample.int(3000,1000,replace=FALSE)
class <- c(rep(TRUE, 500), rep(FALSE, 500))
 pos <- sample.int(1000, 1000)
 random.y <- class[order(pos)]

perfect.x <- 1:3000
perfect.y <- c(rep(TRUE, 1500), rep(FALSE, 1500))

worst.x <- 1:3000
worst.y <- c(rep(FALSE, 1500), rep(TRUE, 1500))

roc.curve(random.x, random.y, file= "randomRocCurve")
roc.curve(perfect.x, perfect.y, file= "perfectRocCurve")
roc.curve(worst.x, worst.y, file= "worstRocCurve") 
}
