% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_algorithm.R
\name{sem}
\alias{sem}
\title{calculate self-embedding model}
\usage{
sem(X, som, grp)
}
\arguments{
\item{X}{A matrix of data.}

\item{som}{A SOM learned X.}

\item{grp}{Graph of manifold equal to som topology.}
}
\value{
SMM object
}
\description{
calculate self-embedding model
}
\examples{
library(TDA)

sphere4d <- sphereUnif(n = 500,d = 3)
spsom <- setSom(4,topo.mode = "sphere")
spgrp <- setGraph(spsom)

spsem <- sem(sphere4d,spsom,spgrp)
}
