% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_class.R
\docType{class}
\name{semR6}
\alias{semR6}
\title{R6 class self-embedding model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
semR6
}
\description{
R6 class self-embedding model
}
\keyword{datasets}
