% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\name{setGraph}
\alias{setGraph}
\title{create graph of manifold from som}
\usage{
setGraph(som, dim = NULL)
}
\arguments{
\item{som}{Corresponding som.}

\item{dim}{embedding dimension.}
}
\description{
create graph of manifold from som
}
