% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_tool.R
\name{setLearningPar}
\alias{setLearningPar}
\title{Sets learning parameters for SOM}
\usage{
setLearningPar(som, alpha = 0.1, sigma = 1, neighbor = 2)
}
\arguments{
\item{som}{target som.}

\item{alpha}{learning rate.}

\item{sigma}{variance of the weights propagation.}

\item{neighbor}{size of neighborhood kernel about the nodes. 1 indicates the neighbors of a node are its one adjacents.}
}
\description{
Sets learning parameters for SOM
}
