% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\name{setPanelSom}
\alias{setPanelSom}
\title{create panel formed som}
\usage{
planesom$new(dim, r, c, weights = NULL, alpha = 0.1, sigma = 1, collect.stats = F)
}
\arguments{
\item{dim}{dimension of input vector}

\item{r}{row of panel topology}

\item{c}{column of panel topology}
}
\description{
create panel formed som
}
