% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_generator.R
\name{setTorusSom}
\alias{setTorusSom}
\title{Title}
\usage{
setTorusSom(dim, r, c, ...)
}
\arguments{
\item{dim}{dimension of input vector.}

\item{r}{row of torusian plane.}

\item{c}{column of torusian plane.}

\item{...}{SOM generater's parameters such as weights = NULL, alpha = 0.1, sigma = 1, neighbor = 1, collect.stats = F.}
}
\value{
torus SOM object.
}
\description{
Title
}
