% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sme_checks}
\alias{sme_checks}
\title{Internal function doing standard checks}
\usage{
sme_checks(
  St3_observed,
  St2_observed,
  St1_observed,
  St3_true,
  St2_true,
  St1_true,
  theta_01,
  theta_02,
  err,
  mu,
  only_2 = TRUE
)
}
\arguments{
\item{St3_observed}{numeric: 0-1}

\item{St2_observed}{numeric: 0-1}

\item{St1_observed}{numeric: 0-1}

\item{St3_true}{numeric: 0-1}

\item{St2_true}{numeric: 0-1}

\item{St1_true}{numeric: 0-1}

\item{theta_01}{numeric}

\item{theta_02}{numeric}

\item{err}{numeric}

\item{mu}{numeric}
}
\value{
error if necessary
}
\description{
Internal function doing standard checks
}
\keyword{internal}
