% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_stata.R
\name{table_stata}
\alias{table_stata}
\title{Table for reporting a conditional logistic model}
\usage{
table_stata(model, exp = TRUE)
}
\arguments{
\item{model}{clogit}

\item{exp}{exponentiate for Odds Ratio or not for Coefficients}
}
\value{

}
\description{
Returns the results of a clogit model. The output matches the one that STATA use.
}
\examples{
#Not run:
#model <- clogit(outcome ~ x1 + x2 + ... + strata(group), data)
#table_stata(model)
}
