% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_uniques.R
\docType{class}
\name{torussom}
\alias{torussom}
\title{R6 generator object of torus formed SOM.}
\format{An \code{R6Class} generator object.}
\usage{
torussom
}
\description{
\code{torussom$new(dim, r, c, weights = NULL, alpha = 0.1, sigma = 1, collect.stats = F)}
}
\keyword{datasets}
