% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_class.R
\docType{class}
\name{tpgrp}
\alias{tpgrp}
\title{R6 generator object of topological graph.}
\format{An R6class generator object.}
\usage{
tpgrp
}
\description{
\code{tpgrp$new(topology, dim = NULL, weights = NULL)}
}
\keyword{datasets}
