% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSOM.R
\docType{class}
\name{tpsom}
\alias{tpsom}
\title{R6 generator object of topological SOM.}
\format{An \code{R6Class} generator object.}
\usage{
tpsom
}
\description{
\code{tpsom$new(dim, topology, weights = NULL, neighbor = 1, alpha = 0.1, sigma = 1, collect.stats = F)}
}
\keyword{datasets}
