% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{variance}
\alias{variance}
\title{Variance}
\usage{
variance(x, margin = 2L)
}
\arguments{
\item{x}{a numeric vector, matrix, or data.frame.}

\item{margin}{In case of matrix or dataframe apply variance by columns margin = 2L or by rows margin = 1L. Default 2.}
}
\value{
the sample variance of the input.
}
\description{
Variance
}
\examples{
example1 <- data.frame("V1" = rep(1, 100), "V2" = 100:1,  "V4" = sample.int(3000,100,replace=FALSE))
example2 <- rep(1, 100)

variance(example1)
variance(example1, margin = 1L)
variance(example2)
}
