% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{visualize.correlation.matrix}
\alias{visualize.correlation.matrix}
\title{Plot correlation matrix}
\usage{
visualize.correlation.matrix(
  cor.df,
  file = "None",
  colors = c("#DEB841", "white", "#267278")
)
}
\arguments{
\item{cor.df}{a data.frame representing correlation matrix}

\item{file}{In case of x been a data.frame or matrix, represent the name of output file for saving mutual information plot.}

\item{colors}{gradiente colors}
}
\value{
Plot correlation matrix
}
\description{
Plot correlation matrix
}
\examples{
df <- data.frame("V1" = sample.int(3000,100,replace=FALSE), "V2" =sample.int(3000,100,replace=FALSE) , "V3" = sample.int(3000,100,replace=FALSE), "V4" = sample.int(3000,100,replace=FALSE),  "V5" = sample.int(3000,100,replace=FALSE), "V6" = sample.int(3000,100,replace=FALSE), "V7" = sample.int(3000,100,replace=FALSE), "V8" = sample.int(3000,100,replace=FALSE))
pearson.cor <- correlation(x = df, method = "Pearson")
visualize.correlation.matrix(cor.df = pearson.cor)
visualize.correlation.matrix(cor.df = pearson.cor, file = "pearsonCor")
visualize.correlation.matrix(cor.df = pearson.cor, file = "pearsonCor", colors = c("Blue", "White", "Green"))
}
