% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{visualize.infotheo.matrix}
\alias{visualize.infotheo.matrix}
\title{Heatmap plot of theory information metrics}
\usage{
visualize.infotheo.matrix(
  df,
  type,
  file = "None",
  colors = c("#DEB841", "white", "#267278")
)
}
\arguments{
\item{df}{Conditional Entropy, Joint Entropy or Mutual Information data.frame or matrix.}

\item{type}{Type of theory information metric. Character}

\item{file}{In case you want to save the plot, file represents output file name.}

\item{colors}{Three colors for the gradient. Default c("#DEB841", "white", "#267278").}
}
\value{
Plot theory information metrics heatmap
}
\description{
Heatmap plot of theory information metrics
}
\examples{
df <- data.frame("V1" = sample.int(2,5,replace=TRUE), "V2" =sample.int(2,5,replace=TRUE) , "V3" = sample.int(2,5,replace=TRUE), "V4" = sample.int(2,5,replace=TRUE),  "V5" = sample.int(2,5,replace=TRUE), "V6" = sample.int(2,5,replace=TRUE), "V7" = sample.int(2,5,replace=TRUE), "V8" = sample.int(2,5,replace=TRUE))
ce <- conditional.entropy(df)
je <- joint.entropy(df)
mi <- mutual.information(df)

visualize.infotheo.matrix(ce, type = "Conditional Entropy")
visualize.infotheo.matrix(ce, type = "Conditional Entropy", file = "conditionalEntropy")
visualize.infotheo.matrix(ce, type = "Conditional Entropy", file = "conditionalEntropy", colors = c("Blue", "White", "Green"))


visualize.infotheo.matrix(je, type = "Joint Entropy")
visualize.infotheo.matrix(je, type = "Joint Entropy", file = "jointEntropy")
visualize.infotheo.matrix(je, type = "Joint Entropy", file = "jointEntropy", colors = c("Blue", "White", "Green"))


visualize.infotheo.matrix(je, type = "Mutual Information")
visualize.infotheo.matrix(je, type = "Mutual Information", file = "mutualInformation")
visualize.infotheo.matrix(je, type = "Mutual Information", file = "mutualInformation", colors = c("Blue", "White", "Green"))
}
