% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{visualize.roc.curve}
\alias{visualize.roc.curve}
\title{Plot Roc curve}
\usage{
visualize.roc.curve(roc.FPR, roc.TPR, auc, file = "None")
}
\arguments{
\item{roc.FPR}{a numerical vector representing false positive rates}

\item{roc.TPR}{a numerical vector representing true positive rates}

\item{auc}{auc value. Use auc() function.}

\item{file}{name of output file for saving roc curve plot.}
}
\value{
Plot roc curve
}
\description{
Plot Roc curve
}
\examples{
random.x <- sample.int(3000,1000,replace=FALSE)
class <- c(rep(TRUE, 500), rep(FALSE, 500))
pos <- sample.int(1000, 1000)
random.y <- class[order(pos)]
roc.curve <- roc.curve(random.x, random.y)
auc <- auc(roc.curve$FPR , roc.curve$TPR )
visualize.roc.curve(roc.curve$FPR , roc.curve$TPR, auc)
visualize.roc.curve(roc.curve$FPR , roc.curve$TPR, auc, file = "randomRocCurve")
}
