% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.df}
\alias{write.df}
\title{Write a csv file}
\usage{
write.df(df, file.path, sep = ";", append = FALSE, header = FALSE)
}
\arguments{
\item{df}{the object to be written, preferably a matrix or data frame.}

\item{file.path}{Name of output file.}

\item{sep}{the field separator string. Values within each row of x are separated by this string. Default ";" .}

\item{append}{If TRUE, the output is appended to the file. If FALSE, any existing file of the name is destroyed. Defatul FALSE.}

\item{header}{if TRUE, name of the columns are written in the first line. Default FALSE.}
}
\description{
Write a csv file
}
\examples{
df = data.frame("V1" = 1:5, "V2" = c("a", "b", "c", "d", "e"), "V3" = c("alto", "bajo", "bajo", "alto", "alto"))
write.df(df, "bench1.txt")
write.df(df, "bench2.txt", header = TRUE)
write.df(df, "bench2.txt", append = TRUE)
write.df(df, "bench2.txt", sep = ";", header = FALSE, append = TRUE)


}
