% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_algorithm.R
\name{e.inner}
\alias{e.inner}
\title{Calculates shift vector on manifold origin.}
\usage{
e.inner(n, grp)
}
\arguments{
\item{n}{numeric of origin.}

\item{grp}{graph of manifold.}
}
\description{
Calculates shift vector on manifold origin.
}
