% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_algorithm.R
\name{p}
\alias{p}
\title{calculate coordinates on manifold}
\usage{
p(X, A, V)
}
\arguments{
\item{X}{Matrix of high dimensional vectors.}

\item{A}{Origin of manifold axis.}

\item{V}{Direction of manifold axis.}
}
\description{
calculate coordinates on manifold
}
