% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\docType{class}
\name{planesom}
\alias{planesom}
\title{R6 generator object of plane formed SOM.}
\format{An \code{R6Class} generator object.}
\usage{
planesom
}
\description{
\code{planesom$new(dim, r, c, weights = NULL, alpha = 0.1, sigma = 1, collect.stats = F)}
}
\keyword{datasets}
