% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_algorithm.R
\name{sem}
\alias{sem}
\title{Calculates self-embedding model.}
\usage{
sem(X, som, grp)
}
\arguments{
\item{X}{a matrix of data.}

\item{som}{a SOM learned X.}

\item{grp}{graph of manifold equal to SOM topology.}
}
\value{
SME object.
}
\description{
Calculates self-embedding model.
}
\examples{
library(TDA)

sphere4d <- sphereUnif(n = 500,d = 3)
spsom <- setSom(4,topo.mode = "sphere")
spgrp <- setGraph(spsom)

spsem <- sem(sphere4d,spsom,spgrp)
}
