% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\name{setGraph}
\alias{setGraph}
\title{Create graph of manifold from SOM.}
\usage{
setGraph(som, dim = NULL)
}
\arguments{
\item{som}{corresponding SOM.}

\item{dim}{embedding dimension.}
}
\description{
Create graph of manifold from SOM.
}
