% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOM_generater.R
\name{setPanelSom}
\alias{setPanelSom}
\title{Creates panel formed SOM.}
\usage{
setPanelSom(dim, r, c)
}
\arguments{
\item{dim}{dimension of input vector.}

\item{r}{row of panel topology.}

\item{c}{column of panel topology.}
}
\description{
Creates panel formed SOM.
}
