% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\name{setSom}
\alias{setSom}
\title{create a type of som}
\usage{
setSom(dim, topo.mode)
}
\arguments{
\item{dim}{Dimension of input vector.}

\item{topo.mode}{Id of typical som topology. Choose from c(grid, plane, circle, sphere, torus, cylinder, cylinder.twocircle).}
}
\value{
Som objsect.
}
\description{
create a type of som
}
\examples{
spsom <- setSom(3,"sphere")
}
