% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\name{setSom}
\alias{setSom}
\title{create a type of som}
\usage{
setSom(dim, topo.mode, weights = NULL, alpha = 0.1, sigma = 1,
  collect.stats = F)
}
\arguments{
\item{dim}{Dimension of input vector.}

\item{topo.mode}{Id of typical som topology. Choose from c(grid, plane, circle, sphere, torus, cylinder, cylinder.twocircle).}

\item{weights}{initialize weights.}

\item{alpha}{learning rate.}

\item{sigma}{variance of the weights propagation.}

\item{collect.stats}{whether rocords learning history.}
}
\value{
Som objsect.
}
\description{
create a type of som
}
\examples{
spsom <- setSom(3,"sphere")
}
