% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_SOM.R
\name{setSom}
\alias{setSom}
\title{Creates a type of SOM.}
\usage{
setSom(dim, topo.mode, weights = NULL, alpha = 0.1, sigma = 1,
  collect.stats = F)
}
\arguments{
\item{dim}{dimension of input vector.}

\item{topo.mode}{type of SOM topology. Choose from c(grid, plane, circle, sphere, torus, cylinder, cylinder.twocircle).}

\item{weights}{initialize weights.}

\item{alpha}{learning rate.}

\item{sigma}{variance of the weights propagation.}

\item{collect.stats}{whether rocords learning history.}
}
\value{
SOM objsect.
}
\description{
Creates a type of SOM.
}
\examples{
spsom <- setSom(3,"sphere")

}
